+++ clients/admin-ui/src/features/organization/index.ts
@@ -1,3 +1,2 @@
 export * from "./constants";
 export * from "./organization.slice";
-export * from "./types";