+++ clients/admin-ui/src/features/taxonomy/types.ts
@@ -1,12 +1,13 @@
-import { FidesBase, FidesKey } from "../common/fides-types";
-
-export interface DataCategory extends FidesBase {
-  parent_key: FidesKey | null;
-}
-
-export interface DataCategoryNode {
+export interface TaxonomyEntityNode {
   value: string;
   label: string;
   description?: string;
-  children: DataCategoryNode[];
+  children: TaxonomyEntityNode[];
+}
+
+export interface TaxonomyEntity {
+  fides_key: string;
+  name?: string;
+  description?: string;
+  parent_key?: string | null;
 }