+++ clients/admin-ui/src/mocks/handlers.ts
@@ -54,7 +54,6 @@ const mockSubjectRequestPreviewResponse = {
   ],
 };
 
-// eslint-disable-next-line import/prefer-default-export
 export const handlers = [
   rest.get<SubjectRequestBody>(
     "http://localhost:8080/api/v1/privacy-request",