+++ src/fidesctl/api/routes/util.py
@@ -5,9 +5,8 @@ from typing import Any, Callable
 from fastapi import HTTPException, status
 
 from fidesctl.api.utils.api_router import APIRouter
-from fidesctl.core.utils import API_PREFIX as _API_PREFIX
 
-API_PREFIX = _API_PREFIX
+API_PREFIX = "/api/v1"
 WEBAPP_DIRECTORY = Path("src/fidesctl/api/build/static")
 WEBAPP_INDEX = WEBAPP_DIRECTORY / "index.html"
 
@@ -56,6 +55,25 @@ def route_requires_okta_connector(func: Callable) -> Callable:
             raise HTTPException(
                 status_code=status.HTTP_400_BAD_REQUEST,
                 detail="Packages not found, ensure aws is included: fidesctl[okta]",
+            )
+        return func(*args, **kwargs)
+
+    return update_wrapper(wrapper_func, func)
+
+
+def route_requires_bigquery_connector(func: Callable) -> Callable:
+    """
+    Function decorator raises a bad request http exception if
+    required modules are not installed for the GCP BigQuery connector
+    """
+
+    def wrapper_func(*args, **kwargs) -> Any:  # type: ignore
+        try:
+            import fidesctl.connectors.bigquery  # pylint: disable=unused-import
+        except ModuleNotFoundError:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Packages not found, ensure BigQuery is included: fidesctl[bigquery]",
             )
         return func(*args, **kwargs)